// src/common/schema.ts
export type Maybe<T> = T | undefined;
// ...
export type CurrencyCode = typeof CurrencyCode[keyof typeof CurrencyCode];
// ...
export type GenericFile = Node & {
	__typename?: 'GenericFile';
	/** A word or phrase to indicate the contents of a file. */
	alt?: Maybe<Scalars['String']>;
	/** A globally-unique identifier. */
	id: Scalars['ID'];
	/** The MIME type of the file. */
	mimeType?: Maybe<Scalars['String']>;
	/** The size of the original file in bytes. */
	originalFileSize?: Maybe<Scalars['Int']>;
	/** The preview image for the file. */
	previewImage?: Maybe<Image>;
	/** The URL of the file. */
	url?: Maybe<Scalars['URL']>;
};
// ...
export type Image = {
	__typename?: 'Image';
	/** A word or phrase to share the nature or contents of an image. */
	altText?: Maybe<Scalars['String']>;
	/** The original height of the image in pixels. Returns `null` if the image is not hosted by Shopify. */
	height?: Maybe<Scalars['Int']>;
	/** A unique identifier for the image. */
	id?: Maybe<Scalars['ID']>;
	/**
	 * The location of the original image as a URL.
	 *
	 * If there are any existing transformations in the original source URL, they will remain and not be stripped.
	 *
	 * @deprecated Use `url` instead
	 */
	originalSrc: Scalars['URL'];
	/**
	 * The location of the image as a URL.
	 * @deprecated Use `url` instead
	 */
	src: Scalars['URL'];
	/**
	 * The location of the transformed image as a URL.
	 *
	 * All transformation arguments are considered "best-effort". If they can be applied to an image, they will be.
	 * Otherwise any transformations which an image type does not support will be ignored.
	 *
	 * @deprecated Use `url(transform:)` instead
	 */
	transformedSrc: Scalars['URL'];
	/**
	 * The location of the image as a URL.
	 *
	 * If no transform options are specified, then the original image will be preserved including any pre-applied transforms.
	 *
	 * All transformation options are considered "best-effort". Any transformation that the original image type doesn't support will be ignored.
	 *
	 * If you need multiple variations of the same image, then you can use [GraphQL aliases](https://graphql.org/learn/queries/#aliases).
	 *
	 */
	url: Scalars['URL'];
	/** The original width of the image in pixels. Returns `null` if the image is not hosted by Shopify. */
	width?: Maybe<Scalars['Int']>;
};
// ...
export type ImageConnection = {
	__typename?: 'ImageConnection';
	/** A list of edges. */
	edges: ImageEdge[];
	/** A list of the nodes contained in ImageEdge. */
	nodes: Image[];
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/** List of supported image content types. */
export const ImageContentType = {
	/** A JPG image. */
	Jpg: 'JPG',
	/** A PNG image. */
	Png: 'PNG',
	/** A WEBP image. */
	Webp: 'WEBP',
} as const;

export type ImageContentType =
	typeof ImageContentType[keyof typeof ImageContentType];
/**
 * An auto-generated type which holds one Image and a cursor during pagination.
 *
 */
export type ImageEdge = {
	__typename?: 'ImageEdge';
	/** A cursor for use in pagination. */
	cursor: Scalars['String'];
	/** The item at the end of ImageEdge. */
	node: Image;
};
// ...
export type MediaImage = Media &
	Node & {
		__typename?: 'MediaImage';
		/** A word or phrase to share the nature or contents of a media. */
		alt?: Maybe<Scalars['String']>;
		/** A globally-unique identifier. */
		id: Scalars['ID'];
		/** The image for the media. */
		image?: Maybe<Image>;
		/** The media content type. */
		mediaContentType: MediaContentType;
		/** The preview image for the media. */
		previewImage?: Maybe<Image>;
	};
// ...
export type Metafield = Node & {
	__typename?: 'Metafield';
	/** The date and time when the storefront metafield was created. */
	createdAt: Scalars['DateTime'];
	/** The description of a metafield. */
	description?: Maybe<Scalars['String']>;
	/** A globally-unique identifier. */
	id: Scalars['ID'];
	/** The key name for a metafield. */
	key: Scalars['String'];
	/** The namespace for a metafield. */
	namespace: Scalars['String'];
	/** The parent object that the metafield belongs to. */
	parentResource: MetafieldParentResource;
	/** Returns a reference object if the metafield definition's type is a resource reference. */
	reference?: Maybe<MetafieldReference>;
	/**
	 * The type name of the metafield.
	 * See the list of [supported types](https://shopify.dev/apps/metafields/definitions/types).
	 *
	 */
	type: Scalars['String'];
	/** The date and time when the storefront metafield was updated. */
	updatedAt: Scalars['DateTime'];
	/** The value of a metafield. */
	value: Scalars['String'];
};
// ...
export type MetafieldParentResource =
	| Article
	| Blog
	| Collection
	| Customer
	| Order
	| Page
	| Product
	| ProductVariant
	| Shop;

/**
 * Returns the resource which is being referred to by a metafield.
 *
 */
export type MetafieldReference =
	| GenericFile
	| MediaImage
	| Page
	| Product
	| ProductVariant
	| Video;
// ...
export type MoneyV2 = {
	__typename?: 'MoneyV2';
	/** Decimal money amount. */
	amount: Scalars['Decimal'];
	/** Currency of the money. */
	currencyCode: CurrencyCode;
};
// ...
export type Node = {
	/** A globally-unique identifier. */
	id: Scalars['ID'];
};
// ...
export type Page = HasMetafields &
	Node &
	OnlineStorePublishable & {
		__typename?: 'Page';
		/** The description of the page, complete with HTML formatting. */
		body: Scalars['HTML'];
		/** Summary of the page body. */
		bodySummary: Scalars['String'];
		/** The timestamp of the page creation. */
		createdAt: Scalars['DateTime'];
		/** A human-friendly unique string for the page automatically generated from its title. */
		handle: Scalars['String'];
		/** A globally-unique identifier. */
		id: Scalars['ID'];
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>;
		/**
		 * A paginated list of metafields associated with the resource.
		 * @deprecated As of 2022-07, the paginated `metafields` field has been repurposed to require a list of metafield identifiers.
		 *
		 */
		metafields: MetafieldConnection;
		/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
		onlineStoreUrl?: Maybe<Scalars['URL']>;
		/** The page's SEO information. */
		seo?: Maybe<Seo>;
		/** The title of the page. */
		title: Scalars['String'];
		/** The timestamp of the latest page update. */
		updatedAt: Scalars['DateTime'];
	};
// ...
export type PageInfo = {
	__typename?: 'PageInfo';
	/** The cursor corresponding to the last node in edges. */
	endCursor?: Maybe<Scalars['String']>;
	/** Whether there are more pages to fetch following the current page. */
	hasNextPage: Scalars['Boolean'];
	/** Whether there are any pages prior to the current page. */
	hasPreviousPage: Scalars['Boolean'];
	/** The cursor corresponding to the first node in edges. */
	startCursor?: Maybe<Scalars['String']>;
};
// ...
export type ProductOption = Node & {
	__typename?: 'ProductOption';
	/** A globally-unique identifier. */
	id: Scalars['ID'];
	/** The product option’s name. */
	name: Scalars['String'];
	/** The corresponding value to the product option name. */
	values: Array<Scalars['String']>;
};

/** The price range of the product. */
export type ProductPriceRange = {
	__typename?: 'ProductPriceRange';
	/** The highest variant's price. */
	maxVariantPrice: MoneyV2;
	/** The lowest variant's price. */
	minVariantPrice: MoneyV2;
};
// ...
export type ProductVariantConnection = {
	__typename?: 'ProductVariantConnection';
	/** A list of edges. */
	edges: ProductVariantEdge[];
	/** A list of the nodes contained in ProductVariantEdge. */
	nodes: ProductVariant[];
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/**
 * An auto-generated type which holds one ProductVariant and a cursor during pagination.
 *
 */
export type ProductVariantEdge = {
	__typename?: 'ProductVariantEdge';
	/** A cursor for use in pagination. */
	cursor: Scalars['String'];
	/** The item at the end of ProductVariantEdge. */
	node: ProductVariant;
};
// ...
export type SelectedOption = {
	__typename?: 'SelectedOption';
	/** The product option’s name. */
	name: Scalars['String'];
	/** The product option’s value. */
	value: Scalars['String'];
};
// ...
export type FullImageFragment = {
	__typename?: 'Image';
	id?: string | undefined;
	width?: number | undefined;
	height?: number | undefined;
	altText?: string | undefined;
	originalSrc: string;
};
// ...
export type FullMetafieldFragment = {
	__typename?: 'Metafield';
	id: string;
	description?: string | undefined;
	key: string;
	namespace: string;
	type: string;
	value: string;
	reference?:
		| {__typename: 'GenericFile'}
		| {
				__typename: 'MediaImage';
				alt?: string | undefined;
				id: string;
				image?:
					| {
							__typename?: 'Image';
							id?: string | undefined;
							width?: number | undefined;
							height?: number | undefined;
							altText?: string | undefined;
							originalSrc: string;
					  }
					| undefined;
		  }
		| {
				__typename: 'Page';
				body: string;
				bodySummary: string;
				id: string;
				title: string;
				handle: string;
		  }
		| {
				__typename: 'Product';
				id: string;
				handle: string;
				availableForSale: boolean;
				title: string;
				productType: string;
				vendor: string;
				description: string;
				descriptionHtml: string;
				totalInventory?: number | undefined;
				options: Array<{
					__typename?: 'ProductOption';
					id: string;
					name: string;
					values: string[];
				}>;
				priceRange: {
					__typename?: 'ProductPriceRange';
					maxVariantPrice: {
						__typename?: 'MoneyV2';
						amount: string;
						currencyCode: CurrencyCode;
					};
					minVariantPrice: {
						__typename?: 'MoneyV2';
						amount: string;
						currencyCode: CurrencyCode;
					};
				};
				variants: {
					__typename?: 'ProductVariantConnection';
					pageInfo: {
						__typename?: 'PageInfo';
						hasNextPage: boolean;
						hasPreviousPage: boolean;
					};
					edges: Array<{
						__typename?: 'ProductVariantEdge';
						cursor: string;
						node: {
							__typename?: 'ProductVariant';
							id: string;
							title: string;
							sku?: string | undefined;
							availableForSale: boolean;
							requiresShipping: boolean;
							quantityAvailable?: number | undefined;
							currentlyNotInStock: boolean;
							selectedOptions: Array<{
								__typename?: 'SelectedOption';
								name: string;
								value: string;
							}>;
							priceV2: {
								__typename?: 'MoneyV2';
								amount: string;
								currencyCode: CurrencyCode;
							};
							compareAtPriceV2?:
								| {
										__typename?: 'MoneyV2';
										amount: string;
										currencyCode: CurrencyCode;
								  }
								| undefined;
							image?:
								| {
										__typename?: 'Image';
										id?: string | undefined;
										width?: number | undefined;
										height?: number | undefined;
										altText?: string | undefined;
										originalSrc: string;
								  }
								| undefined;
						};
					}>;
				};
				images: {
					__typename?: 'ImageConnection';
					pageInfo: {
						__typename?: 'PageInfo';
						hasNextPage: boolean;
						hasPreviousPage: boolean;
					};
					edges: Array<{
						__typename?: 'ImageEdge';
						cursor: string;
						node: {
							__typename?: 'Image';
							id?: string | undefined;
							width?: number | undefined;
							height?: number | undefined;
							altText?: string | undefined;
							originalSrc: string;
						};
					}>;
				};
		  }
		| {
				__typename: 'ProductVariant';
				id: string;
				title: string;
				sku?: string | undefined;
				availableForSale: boolean;
				requiresShipping: boolean;
				quantityAvailable?: number | undefined;
				currentlyNotInStock: boolean;
				product: {
					__typename?: 'Product';
					id: string;
					title: string;
					handle: string;
					productType: string;
					vendor: string;
				};
				selectedOptions: Array<{
					__typename?: 'SelectedOption';
					name: string;
					value: string;
				}>;
				priceV2: {
					__typename?: 'MoneyV2';
					amount: string;
					currencyCode: CurrencyCode;
				};
				compareAtPriceV2?:
					| {
							__typename?: 'MoneyV2';
							amount: string;
							currencyCode: CurrencyCode;
					  }
					| undefined;
				image?:
					| {
							__typename?: 'Image';
							id?: string | undefined;
							width?: number | undefined;
							height?: number | undefined;
							altText?: string | undefined;
							originalSrc: string;
					  }
					| undefined;
		  }
		| {__typename: 'Video'}
		| undefined;
};